project-root/backend/requirements.txt
fastapi[security]==0.104.1
uvicorn==0.24.0.post1
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
alembic==1.12.1
pydantic==2.5.2
passlib[bcrypt]==1.7.4
python-jose[cryptography]==3.3.0
prometheus_client==0.19.0
apscheduler==3.10.4
slowapi==0.1.9